+++ build.common.gradle
@@ -59,7 +59,7 @@ android {
         //noinspection ExpiredTargetSdkVersion
         targetSdkVersion 28
 
-
+        //multiDexEnabled true
         /**
          * We keep the versionCode and versionName of robot controller applications in sync with
          * the master information published in the AndroidManifest.xml file of the FtcRobotController
